+++ libcrux-ml-kem/src/ind_cca.rs
@@ -5,9 +5,9 @@ use crate::{
     constants::{CPA_PKE_KEY_GENERATION_SEED_SIZE, H_DIGEST_SIZE, SHARED_SECRET_SIZE},
     hash_functions::Hash,
     polynomial::PolynomialRingElement,
-    ind_cpa::serialize_public_key,
+    ind_cpa::{unpacked::*,serialize_public_key},
     serialize::deserialize_ring_elements_reduced,
-    types::{unpacked::*, *},
+    types::*,
     utils::into_padded_array,
     vector::Operations,
 };
