+++ libcrux-ml-kem/src/mlkem1024.rs
@@ -2,8 +2,9 @@
 
 use super::{
     constants::*,
+    ind_cpa::unpacked::*,
     ind_cca::*,
-    types::{unpacked::*, *},
+    types::*,
     vector::traits::VectorType,
     *,
 };
