+++ libcrux-ml-kem/src/mlkem512.rs
@@ -1,8 +1,9 @@
 //! ML-KEM 512
 use super::{
     constants::*,
+    ind_cpa::unpacked::*,
     ind_cca::*,
-    types::{unpacked::*, *},
+    types::*,
     vector::traits::VectorType,
     *,
 };
